Add crude search engine
Index needs to be created; can't find a program I wrote to do just that,
will have to wait until I find it.  Or not, as it's pretty trivial actually.
In any case, the index is loaded via AJAX only if needed, so there's very
little overhead if nobody uses the search feature (or if someone uses it by
loading a page from a search hit).

A nice enhancement would be adding all indexed words to a tree to perform a
prefix search -- and for each submatch, perform a whole search; then,
perform an union of all of those result sets.

Another enhancement would be porting over Peter Norvig's spell checker to
find similar words in the index and perform searches on top of the
suggestions as well.

What would be really nice is actually a way to find which document is more
relevant.  Currently, only the top N hits are shown, in an arbitrary order.
This is clearly not optimal.  Either order it better, or show N items, and
if there are N+ items, add another popup item to show a larger window with
all hits.  Or a combination of both.